Move ModulesDirectory to another class
