+++ lib/npmdc/checker.rb
@@ -4,6 +4,7 @@ require 'forwardable'
 require 'semantic_range'
 require 'npmdc/formatter'
 require 'npmdc/errors'
+require 'npmdc/modules_directory'
 
 module Npmdc
   class Checker
@@ -60,50 +61,6 @@ module Npmdc
                      'Everything is ok.'
 
       output(result_stats, :success)
-    end
-
-    class ModulesDirectory
-      attr_reader :path
-
-      def initialize(path)
-        @path = path
-      end
-
-      def basename
-        File.basename(path)
-      end
-
-      def scoped?
-        basename.start_with?('@')
-      end
-
-      def files
-        Dir.glob("#{path}/*").map { |file_path| self.class.new(file_path) }
-      end
-
-      def directories
-        files.select(&:directory?)
-      end
-
-      def valid_directories
-        directories.select { |d| d.package_json_file.exists? || d.scoped? }
-      end
-
-      def package_json_file
-        self.class.new(File.join(path, 'package.json'))
-      end
-
-      def file?
-        File.file?(path)
-      end
-
-      def directory?
-        File.directory?(path)
-      end
-
-      def exists?
-        File.file?(path)
-      end
     end
 
     def modules_directory_path
